requirements.txt
# #Langchain Core
# langchain
# langchain-core

# #Google Gemini Integration
# langchain-google-genai
# google-generativeai
# langchain[google-genai]

# #hungging Face Ientegration
# langchain-huggingface
# transformers
# huggingface_hub

# #environment variable Mangement
# python-dotenv

# #Machine Learning Utilities
# scikit-learn
# numpy
# pandas

# # Core LLM frameworks
# langchain-core>=1.2.5
# langchain-google-genai>=4.2.0
# langchain-huggingface>=0.1.0

# # HuggingFace libraries
# huggingface-hub>=0.20.0
# transformers>=4.30.0
# sentence-transformers>=2.2.0

# # Deep Learning frameworks
# torch>=2.0.0
# torchvision>=0.15.0
# torchaudio>=2.0.0

# # API clients
# google-genai>=1.56.0

# # Utilities
# python-dotenv>=1.0.0
# requests>=2.28.0
# pydantic>=2.0.0

# # Optional but recommended
# accelerate>=0.20.0
# sentencepiece>=0.1.99
# protobuf>=3.20.0

# Core Langchain
langchain>=0.3.0
langchain-core>=1.2.5

# Google Gemini Integration
langchain-google-genai>=4.2.0
google-genai>=1.56.0

# HuggingFace Integration
langchain-huggingface>=0.1.0
huggingface-hub>=0.20.0
transformers>=4.30.0
sentence-transformers>=2.2.0

# Deep Learning Frameworks
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Environment Variable Management
python-dotenv>=1.0.0

# Machine Learning Utilities
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# Core Utilities
requests>=2.28.0
pydantic>=2.0.0

# Performance & Optimization
accelerate>=0.20.0
sentencepiece>=0.1.99
protobuf>=3.20.0